requirements.txt
fastapi
uvicorn[standard]
python-dotenv
pydantic
sqlalchemy>=2.0
psycopg2-binary
pgvector
openai>=1.0.0
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
beautifulsoup4
httpx
